# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medora academic API",
    "version": "0.1.0",
    "description": "Development contract. Domain creation inputs generated from runtime Zod schemas. Learning, identity and action-specific contracts remain documented separately."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/documents": {
      "get": {
        "summary": "List authorized documents",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped paginated records"
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Role forbidden"
          }
        }
      },
      "post": {
        "summary": "Create documents",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentsInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created record"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
    "/evidence": {
      "get": {
        "summary": "List authorized evidence",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped paginated records"
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Role forbidden"
          }
        }
      },
      "post": {
        "summary": "Create evidence",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created record"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "medora_session"
      }
    },
    "schemas": {
...
      "DocumentsInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 250
          },
          "mime": {
            "type": "string",
            "enum": [
              "application/pdf",
              "image/png",
              "image/jpeg"
            ]
          },
          "content_base64": {
            "type": "string",
            "maxLength": 7000000
          }
        },
        "required": [
          "name",
          "mime",
          "content_base64"
        ],
        "additionalProperties": false
      },
      "EvidenceInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 250
          },
          "period": {
            "type": "string",
            "minLength": 1,
            "maxLength": 250
          }
        },
        "required": [
          "title",
          "period"
        ],
        "additionalProperties": false
      }
    }
  }
}