# packages/contracts/types.ts
// Generated by scripts/contracts.ts. Do not edit.
export type ApplicationsInput = { "name": string; "email": string; "external_ref": string; "pool_id": string; "department": string; };
export type SessionsInput = { "title": string; "department": string; "cohort": string; "faculty_id": string; "competency_id": string; "room": string; "category": "Theory" | "Practical" | "Posting"; "starts_at": string; "ends_at": string; };
export type CompetenciesInput = { "code": string; "title": string; "subject": string; "curriculum_version": string; "source": string; };
export type AssessmentsInput = { "title": string; "department": string; "cohort": string; "competency_id": string; "max_marks": number; "examiner_id": string; "rubric": string; };
export type LogbookInput = { "competency_id": string; "supervisor_id": string; "activity": string; "activity_date": string; "reflection": string; "posting": string; };
export type InvoicesInput = { "student_id": string; "description": string; "amount_minor": number; "due_date": string; "fee_version": string; };
export type ContentInput = { "slug": string; "title": string; "body": string; "kind": "Page" | "Notice" | "Event" | "Disclosure"; "language": "en" | "bn"; "review_date": string; "category"?: "General" | "Admissions" | "Academic" | "Examinations" | "Student services" | "Research" | "Recruitment" | "Tender"; "issue_date"?: string | null; "reference"?: string; "available_on"?: string | null; "archive_on"?: string | null; };
export type TicketsInput = { "title": string; "description": string; "confidential": boolean; "severity": "Normal" | "High" | "Urgent"; "due_date": string; };
export type NoticesInput = { "title": string; "body": string; "audience": "all" | "student" | "faculty" | "finance" | "registrar"; };
export type MastersInput = { "kind": "Campus" | "Department" | "Programme" | "Batch" | "Hospital"; "code": string; "name": string; };
export type PoliciesInput = { "name": string; "kind": "Attendance" | "Academic" | "Fees"; "effective_date": string; "config": Record<string, unknown>; };
export type CorrectionsInput = { "session_id": string; "student_id": string; "proposed_status": "Present" | "Absent" | "Excused"; "reason": string; };
export type RefundsInput = { "payment_id": string; "amount_minor": number; "reason": string; };
export type DocumentsInput = { "name": string; "mime": "application/pdf" | "image/png" | "image/jpeg"; "content_base64": string; };
export type EvidenceInput = { "title": string; "period": string; };
export type ApiErrorEnvelope = { error: {code:string;message:string;correlation_id?:string} };